moving make moderator intent handler to separate file
